@@ -18,14 +18,15 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  objectbox: ^0.11.0
-  objectbox_flutter_libs: ^0.11.0
-  path_provider: ^1.6.27
+  objectbox: ^1.5.0
+  objectbox_flutter_libs: ^1.5.0
+  path_provider: ^2.0.10
+  flutter_lints: ^2.0.0
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -35,8 +36,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.11.1
-  objectbox_generator: ^0.11.0
+  build_runner: ^2.1.11
+  objectbox_generator: ^1.5.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
