@@ -1,12 +1,15 @@
 import 'dart:async';
-import '../controller/contact_controller.dart';
+import 'package:contactlist/main.dart';
+
 import '../model/contact_model.dart';
 import '../view/widgets/contact_list_widget.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:flutter/material.dart';
 
 class ContactListView extends StatefulWidget {
+  const ContactListView({Key? key}) : super(key: key);
+
   @override
+  // ignore: library_private_types_in_public_api
   _ContactListViewState createState() => _ContactListViewState();
 }
 
@@ -17,27 +20,25 @@ class _ContactListViewState extends State<ContactListView> {
 
   final TextEditingController _nameControl = TextEditingController();
   final TextEditingController _emailControl = TextEditingController();
-  ContactController _contactController;
-  Stream<List<ContactModel>> _streamContact;
-  String _name;
-  String _email;
+  // final ContactController _contactController;
+  // late final Stream<List<ContactModel>> _streamContact;
+  final _listContactStream = StreamController<List<ContactModel>>(sync: true);
+  String? _name;
+  String? _email;
 
   @override
   void initState() {
     super.initState();
-    getApplicationDocumentsDirectory().then((dir) {
-      _contactController = ContactController(dir);
-      _streamContact = _listController.stream;
-      _listController.add(_contactController.allContacts);
-      _listController.addStream(_contactController.queryStream);
-
-      setState(() {});
-    });
+    // _streamContact = _listController.stream as Stream<List<ContactModel>>;
+    // _listController.add(_contactController.allContacts);
+    // _listController.addStream(_contactController.queryStream);
+    _listContactStream.addStream(contactController.queryStream);
+    setState(() {});
   }
 
   @override
   void dispose() {
-    _contactController.dispose();
+    // _contactController.dispose();
     _listController.close();
     _nameControl.dispose();
     _emailControl.dispose();
@@ -48,7 +49,7 @@ class _ContactListViewState extends State<ContactListView> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Lista de Contatos'),
+        title: const Text('Lista de Contatos'),
         centerTitle: true,
       ),
       body: SafeArea(
@@ -58,7 +59,7 @@ class _ContactListViewState extends State<ContactListView> {
             children: [
               SizedBox(
                 width: MediaQuery.of(context).size.width,
-                height: 80,
+                height: MediaQuery.of(context).size.height * 0.15,
                 child: Column(
                   mainAxisSize: MainAxisSize.min,
                   children: [
@@ -71,23 +72,22 @@ class _ContactListViewState extends State<ContactListView> {
                             child: TextFormField(
                               controller: _nameControl,
                               validator: (name) =>
-                                  name.isEmpty ? 'Nome inválido' : null,
-                              decoration: InputDecoration(
+                                  name!.isEmpty ? 'Nome inválido' : null,
+                              decoration: const InputDecoration(
                                   labelText: 'Nome', filled: true),
                               onSaved: (name) => _name = name,
                               keyboardType: TextInputType.name,
                               textCapitalization: TextCapitalization.words,
                             ),
                           ),
-                          Padding(
-                              padding:
-                                  const EdgeInsets.symmetric(horizontal: 4)),
+                          const Padding(
+                              padding: EdgeInsets.symmetric(horizontal: 4)),
                           Expanded(
                             child: TextFormField(
                               controller: _emailControl,
                               validator: (email) =>
-                                  email.isEmpty ? 'E-mail inválido' : null,
-                              decoration: InputDecoration(
+                                  email!.isEmpty ? 'E-mail inválido' : null,
+                              decoration: const InputDecoration(
                                   labelText: 'E-mail', filled: true),
                               onSaved: (email) => _email = email,
                               keyboardType: TextInputType.emailAddress,
@@ -104,19 +104,17 @@ class _ContactListViewState extends State<ContactListView> {
                 children: [
                   SizedBox(
                     width: MediaQuery.of(context).size.width / 2,
-                    child: RaisedButton(
+                    child: ElevatedButton(
                       onPressed: () {
-                        if (_formKey.currentState.validate()) {
-                          _formKey.currentState.save();
-                          _contactController.addContact(ContactModel.construct(
+                        if (_formKey.currentState!.validate()) {
+                          _formKey.currentState?.save();
+                          contactController.addContact(ContactModel.construct(
                               name: _name, email: _email));
                         }
                         _nameControl.clear();
                         _emailControl.clear();
                       },
-                      shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(20)),
-                      child: Text(
+                      child: const Text(
                         'Adicionar',
                         style: TextStyle(fontSize: 16),
                       ),
@@ -126,39 +124,42 @@ class _ContactListViewState extends State<ContactListView> {
               ),
               Expanded(
                 child: StreamBuilder<List<ContactModel>>(
-                  stream: _streamContact,
+                  stream: _listContactStream.stream,
                   builder: (_, snapshot) {
                     if (!snapshot.hasData) {
-                      return Center(
+                      return const Center(
                         child: CircularProgressIndicator(),
                       );
-                    } else if (snapshot.data.isEmpty) {
-                      return Center(
+                    } else if (snapshot.data!.isEmpty) {
+                      return const Center(
                         child: Text('Não há contatos em sua lista.'),
                       );
                     } else {
                       return ListView.builder(
-                          itemCount: snapshot.data.length,
-                          itemBuilder: (_, index) {
-                            final contact = snapshot.data[index];
-                            return ContactListWidget(
-                              contact: contact,
-                              onPressedEdit: () {
-                                if (_formKey.currentState.validate()) {
-                                  _formKey.currentState.save();
-                                  _contactController.addContact(
-                                      ContactModel.construct(
-                                          id: contact.id,
-                                          name: _name,
-                                          email: _email));
-                                }
-                                _nameControl.clear();
-                                _emailControl.clear();
-                              },
-                              onPressedDelete: () =>
-                                  _contactController.removeContact(contact),
-                            );
-                          });
+                        itemCount: snapshot.data?.length,
+                        itemBuilder: (_, index) {
+                          final contact = snapshot.data?[index];
+                          return ContactListWidget(
+                            contact: contact,
+                            onPressedEdit: () {
+                              if (_formKey.currentState!.validate()) {
+                                _formKey.currentState?.save();
+                                contactController.addContact(
+                                  ContactModel.construct(
+                                    id: contact?.id,
+                                    name: _name,
+                                    email: _email,
+                                  ),
+                                );
+                              }
+                              _nameControl.clear();
+                              _emailControl.clear();
+                            },
+                            onPressedDelete: () =>
+                                contactController.removeContact(contact!),
+                          );
+                        },
+                      );
                     }
                   },
                 ),
