@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'view/contact_view.dart';
 
 class App extends StatelessWidget {
+  const App({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -11,7 +13,7 @@ class App extends StatelessWidget {
       darkTheme: ThemeData(
           brightness: Brightness.dark,
           primaryColor: Colors.deepPurpleAccent[700]),
-      home: ContactListView(),
+      home: const ContactListView(),
     );
   }
 }
