@@ -1,82 +1,133 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// This code was generated by ObjectBox. To update it run the generator again:
+// With a Flutter package, run `flutter pub run build_runner build`.
+// With a Dart package, run `dart run build_runner build`.
+// See also https://docs.objectbox.io/getting-started#generate-objectbox-code
 
-// Currently loading model from "JSON" which always encodes with double quotes
-// ignore_for_file: prefer_single_quotes
+// ignore_for_file: camel_case_types
 
 import 'dart:typed_data';
 
+// ignore: depend_on_referenced_packages
+import 'package:flat_buffers/flat_buffers.dart' as fb;
+import 'package:objectbox/internal.dart'; // generated code can access "internal" functionality
 import 'package:objectbox/objectbox.dart';
-import 'package:objectbox/flatbuffers/flat_buffers.dart' as fb;
+import 'package:objectbox_flutter_libs/objectbox_flutter_libs.dart';
+
+import 'model/contact_model.dart';
+
 export 'package:objectbox/objectbox.dart'; // so that callers only have to import this file
-import 'package:objectbox/internal.dart';
 
-import 'model/contact_model.dart'; // generated code can access "internal" functionality
+final _entities = <ModelEntity>[
+  ModelEntity(
+      id: const IdUid(1, 2761970928210775872),
+      name: 'ContactModel',
+      lastPropertyId: const IdUid(3, 7689849335879901880),
+      flags: 0,
+      properties: <ModelProperty>[
+        ModelProperty(
+            id: const IdUid(1, 7172456119015303439),
+            name: 'id',
+            type: 6,
+            flags: 1),
+        ModelProperty(
+            id: const IdUid(2, 1335065517310788035),
+            name: 'name',
+            type: 9,
+            flags: 0),
+        ModelProperty(
+            id: const IdUid(3, 7689849335879901880),
+            name: 'email',
+            type: 9,
+            flags: 0)
+      ],
+      relations: <ModelRelation>[],
+      backlinks: <ModelBacklink>[])
+];
 
+/// Open an ObjectBox store with the model declared in this file.
+Future<Store> openStore(
+        {String? directory,
+        int? maxDBSizeInKB,
+        int? fileMode,
+        int? maxReaders,
+        bool queriesCaseSensitiveDefault = true,
+        String? macosApplicationGroup}) async =>
+    Store(getObjectBoxModel(),
+        directory: directory ?? (await defaultStoreDirectory()).path,
+        maxDBSizeInKB: maxDBSizeInKB,
+        fileMode: fileMode,
+        maxReaders: maxReaders,
+        queriesCaseSensitiveDefault: queriesCaseSensitiveDefault,
+        macosApplicationGroup: macosApplicationGroup);
 
+/// ObjectBox model definition, pass it to [Store] - Store(getObjectBoxModel())
 ModelDefinition getObjectBoxModel() {
-  final model = ModelInfo.fromMap({
-    "entities": [
-      {
-        "id": "1:5610599817304566832",
-        "lastPropertyId": "3:9073858309807067231",
-        "name": "ContactModel",
-        "properties": [
-          {"id": "1:5687740588200089885", "name": "id", "type": 6, "flags": 1},
-          {"id": "2:4619961722547692663", "name": "name", "type": 9},
-          {"id": "3:9073858309807067231", "name": "email", "type": 9}
-        ],
-        "relations": [],
-        "backlinks": []
-      }
-    ],
-    "lastEntityId": "1:5610599817304566832",
-    "lastIndexId": "0:0",
-    "lastRelationId": "0:0",
-    "lastSequenceId": "0:0",
-    "modelVersion": 5
-  }, check: false);
+  final model = ModelInfo(
+      entities: _entities,
+      lastEntityId: const IdUid(1, 2761970928210775872),
+      lastIndexId: const IdUid(0, 0),
+      lastRelationId: const IdUid(0, 0),
+      lastSequenceId: const IdUid(0, 0),
+      retiredEntityUids: const [],
+      retiredIndexUids: const [],
+      retiredPropertyUids: const [],
+      retiredRelationUids: const [],
+      modelVersion: 5,
+      modelVersionParserMinimum: 5,
+      version: 1);
 
-  final bindings = <Type, EntityDefinition>{};
-  bindings[ContactModel] = EntityDefinition<ContactModel>(
-      model: model.getEntityByUid(5610599817304566832),
-      toOneRelations: (ContactModel object) => [],
-      toManyRelations: (ContactModel object) => {},
-      getId: (ContactModel object) => object.id,
-      setId: (ContactModel object, int id) {
-        object.id = id;
-      },
-      objectToFB: (ContactModel object, fb.Builder fbb) {
-        final offsetname =
-            object.name == null ? null : fbb.writeString(object.name);
-        final offsetemail =
-            object.email == null ? null : fbb.writeString(object.email);
-        fbb.startTable();
-        fbb.addInt64(0, object.id ?? 0);
-        fbb.addOffset(1, offsetname);
-        fbb.addOffset(2, offsetemail);
-        fbb.finish(fbb.endTable());
-        return object.id ?? 0;
-      },
-      objectFromFB: (Store store, Uint8List fbData) {
-        final buffer = fb.BufferContext.fromBytes(fbData);
-        final rootOffset = buffer.derefObject(0);
+  final bindings = <Type, EntityDefinition>{
+    ContactModel: EntityDefinition<ContactModel>(
+        model: _entities[0],
+        toOneRelations: (ContactModel object) => [],
+        toManyRelations: (ContactModel object) => {},
+        getId: (ContactModel object) => object.id,
+        setId: (ContactModel object, int id) {
+          object.id = id;
+        },
+        objectToFB: (ContactModel object, fb.Builder fbb) {
+          final nameOffset =
+              object.name == null ? null : fbb.writeString(object.name!);
+          final emailOffset =
+              object.email == null ? null : fbb.writeString(object.email!);
+          fbb.startTable(4);
+          fbb.addInt64(0, object.id ?? 0);
+          fbb.addOffset(1, nameOffset);
+          fbb.addOffset(2, emailOffset);
+          fbb.finish(fbb.endTable());
+          return object.id ?? 0;
+        },
+        objectFromFB: (Store store, ByteData fbData) {
+          final buffer = fb.BufferContext(fbData);
+          final rootOffset = buffer.derefObject(0);
 
-        final object = ContactModel();
-        object.id = fb.Int64Reader().vTableGet(buffer, rootOffset, 4);
-        object.name = fb.StringReader().vTableGet(buffer, rootOffset, 6);
-        object.email = fb.StringReader().vTableGet(buffer, rootOffset, 8);
-        return object;
-      });
+          final object = ContactModel()
+            ..id =
+                const fb.Int64Reader().vTableGetNullable(buffer, rootOffset, 4)
+            ..name = const fb.StringReader(asciiOptimization: true)
+                .vTableGetNullable(buffer, rootOffset, 6)
+            ..email = const fb.StringReader(asciiOptimization: true)
+                .vTableGetNullable(buffer, rootOffset, 8);
+
+          return object;
+        })
+  };
 
   return ModelDefinition(model, bindings);
 }
 
-// ignore: camel_case_types
+/// [ContactModel] entity fields to define ObjectBox queries.
 class ContactModel_ {
+  /// see [ContactModel.id]
   static final id =
-      QueryIntegerProperty(entityId: 1, propertyId: 1, obxType: 6);
+      QueryIntegerProperty<ContactModel>(_entities[0].properties[0]);
+
+  /// see [ContactModel.name]
   static final name =
-      QueryStringProperty(entityId: 1, propertyId: 2, obxType: 9);
+      QueryStringProperty<ContactModel>(_entities[0].properties[1]);
+
+  /// see [ContactModel.email]
   static final email =
-      QueryStringProperty(entityId: 1, propertyId: 3, obxType: 9);
+      QueryStringProperty<ContactModel>(_entities[0].properties[2]);
 }
